requirements.txt
langchain==0.1.9
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
streamlit==1.31.0
nltk==3.8.1
tiktoken==0.5.2
chromadb==0.4.20
wikipedia==1.4.0
requests==2.31.0
pypdf==3.17.1
pdfminer.six==20221105
google-generativeai==0.3.1 